Attempt to describe OpenOCD
This is my attempt to implement the OpenOCD TODO item in the 'Tooling'
section of Chapter 1.

I did move the OpenOCD section so that it comes after GDB since I
thought it made more sense to describe it after the user has already
been introduced to GDB as a debugger since it is OpenOCD which bridges
the gap between GDB and the ST-Link debugger hardware.

Feedback welcome!