Have mdBook not wrap code snippets in "fn main() {}"
I noticed that when I tried to copy a code snippet into the clipboard
that it had some extra code text that I didn't see on the webpage. It
had wrapped the code snippet in a "fn main()" declaration. The code
window on the page also contained arrows in the upper-right corner for
expanding the code so that this hidden declaration could be seen. These
snippets aren't meant to be working samples so treating them as code to
be contained in main() often doesn't make sense. Changing the language
from "rust" to "rust,ignore" on these snippets causes mdBook to not
perform this wrapping.

For example before the following would be copied to clipboard:
```rust
fn main() {
// Exception handler for the SysTick (System Timer) exception
fn SysTick() {
    // ..
}
}
```

After this change, we get the following copied to the clipboard:
```rust
// Exception handler for the SysTick (System Timer) exception
fn SysTick() {
    // ..
}
```