Move "last_state = state" assignment
I believe that I have found a bug in the samples found in the
concurrency chapter where the value of the `last_state` variable
should be updated on each loop iteration but it is only updated
on the detection of the first leading edge. For example:

    loop {
        let state = read_signal_level();
        if state && !last_state {
            last_state = state;
            // DANGER - Not actually safe! Could cause data races.
            unsafe { COUNTER += 1 };
        }
    }

I think that loop should actually be written as:

    loop {
        let state = read_signal_level();
        if state && !last_state {
            // DANGER - Not actually safe! Could cause data races.
            unsafe { COUNTER += 1 };
        }
        last_state = state;
    }

As the code exists today, I think it would detect the first leading
edge, set last_state to true and then never be updated again which
would also mean that it never detects another leading edge.