collections: Fix example allocator
The example bump-pointer allocator doesn't take into account the size of
the allocation, just the alignment. As a result it'll work fine for small
allocations where alignment >= size (e.g. a single int), but it results in
overlapping allocations for anything larger.

Also, change the alignment calculation to find the aligment using the
add-then-mask method rather than screwing around with remainders.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>