Make glossary more linkable and add more detail
While writing a blog post I found that I wanted to be able to link to a
common definition of certain terms but I was not able to find one.

This changes makes the glossary up from sub-headings so that each item
can be linked to directly. I have also filled out some more detail on
the areas I'm particularly interested in (BSP, HAL and PAC).