Use mdbook-i18n-helpers to support translations
This PR expands the publishing step to also publish a (very partial)
Danish translation of the book.

The infrastructure uses the GNU Gettext tools, as well as a mdbook
plugin:

    sudo apt install gettext
    cargo install mdbook-i18n-helpers

The translation lives in po/da.po, which was created with

    MDBOOK_OUTPUT='{"xgettext": {"pot-file": "messages.pot"}}' mdbook build -d po
    msginit -l da -i po/messages.pot -o po/da.po

After creating the da.po file, the translation can be served locally
with:

    MDBOOK_BOOK__LANGUAGE=da mdbook serve

The publication flow on GitHub has been expanded to publish
translations into subdirectories named after the two-letter ISO_639-1
language codes. The English source text stays the same.

Links to the new translations are added manually in the index.hbs
theme file[1]. I imported the one from mdbook and added a very simple
language picker in the top right corner.

[1]: https://raw.githubusercontent.com/rust-lang/mdBook/master/src/theme/index.hbs